+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* **2015-03-17**: Added support for marketplace hints in basket and article details.
 * **2015-03-17**: [BC-BREAK] Removed module setting, replaced by own module configuration
 * **2015-03-17**: Module configuration in own admin area 
 * **2015-03-13**: Admin area for bebado module setting